Isolate some fixture related db code
